BUGFIX: fixed name generator for long commands
Job names are created from the commands that will be run, including the
command arguments. This causes problems for long commands (like command
with several arguments) with the PBS handler because the job name is
used to name the output file. If the output filename is too long, it can
the job will error.

Now, if the generated job name is longer than 40 characters, an md5 sum
of the command is used instead.